Made property search server side, got rid of API route.
